+++ pandas/tests/reductions/test_reductions.py
@@ -159,6 +159,14 @@ class TestReductions:
         expected = df[expected_col].rename(None)
         tm.assert_series_equal(result, expected)
 
+    def test_numpy_reduction_with_tz_aware_dtype(self, tz_aware_fixture):
+        # GH 15552
+        tz = tz_aware_fixture
+        arg = pd.to_datetime(["2019"]).tz_localize(tz)
+        expected = Series(arg)
+        result = np.minimum(expected, expected)
+        tm.assert_series_equal(result, expected)
+
 
 class TestIndexReductions:
     # Note: the name TestIndexReductions indicates these tests
